Extract file
